+++ core/blocks.py
@@ -510,3 +510,10 @@ class TopicPageCardBlock(blocks.StructBlock):
     link_url = blocks.CharBlock()  # not a URL block to allow relative links
     image = ImageChooserBlock(required=False)  #  alt text lives on the custom Image class
     description = blocks.TextBlock()
+
+
+class ColumnsBlock(blocks.StructBlock):
+    title = blocks.CharBlock(required=False, max_length=255, label='Title')
+    image = ImageChooserBlock(required=False, label='Hero Image')
+    description = blocks.RichTextBlock(features=RICHTEXT_FEATURES__REDUCED, required=False, label='Description')
+    link = blocks.URLBlock(required=False, label='Title link')